+++ Game/src/mge/util/BiomeHandler.cpp
@@ -46,13 +46,13 @@ int BiomeHandler::GetTotalActiveInstances()
 }
 
 
-void BiomeHandler::GivePreset(int biomeID, size_t index, GameObject* preset) 
+void BiomeHandler::GivePreset(size_t biomeID, size_t index, GameObject* preset) 
 {
 	if (biomeID >= _biomes.size())return;
 	return _biomes[biomeID]->GivePreset(index, preset);
 }
 
-GameObject* BiomeHandler::TakePreset(int biomeID, size_t index) 
+GameObject* BiomeHandler::TakePreset(size_t biomeID, size_t index) 
 {
 	if (biomeID >= _biomes.size())return nullptr;
 	return _biomes[biomeID]->TakePreset(index);
